--- src/presenter/points.js
@@ -1,16 +1,0 @@
-
-export default class points {
-  constructor() {
-  }
-
-}
-
-
-_handleSortTypeChange() {
-}
-
-_renderSort() {
-}
-
-_renderTask() {
-}
